test: FeaturedFairs
